requirements.txt
fastapi
uvicorn
jinja2
python-multipart
beautifulsoup4
playwright
openai>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
sqlalchemy>=2.0.0,<3.0.0
pytest
